requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-text-splitters
langchain-huggingface
sentence-transformers
langchain-groq
pypdf
faiss-cpu
python-dotenv
